requirements.txt
streamlit==1.27.2
flask==2.3.2
PyPDF2==3.0.1
python-docx==0.8.11
spacy==3.5.2
deep-translator==1.9.1
pandas==2.0.3
numpy==1.24.3
reportlab==3.6.12
protobuf==3.20.0